Add tab history to remain functional even with closed tabs
`tabHistory` is a doubly linked list with random access using tab id. A
dummy start is placed at the beginning of the list. Only visited tabs
are added to the list (`tabs.onActivated` listener), and as soon as they
are closed, they're removed from the list (`tabs.onRemoved` listener).

Helper function `validateHistory()` helps out a lot for debugging.